msm7x30: Camera HAL for HTC Sense devices (3/2)
This adds an option to the camera to "force" the preview. When using a
wrapper HAL for a GB lib camera, this will fix preview not being
available until a picture is taken.
<string name="pref_camera_force_preview_default">on</string> in a
string.xml overlay will activate the option, which has also been
designed to support being added as a menu selection for private testing

Change-Id: I7c0a297690d5b191dbaf9fd6ff61548be58c056c