Camera: HDR: Read parameters for hardware HDR support
On mako (and soon find5), HDR mode is a scene-mode. Software HDR
uses Util.isCameraHdrSupported to check whether or not hardware
already supports it, which should also set software HDR flag.
However, this method is not called soon enough to set this flag
on time, so we call it before checking for software HDR status.

Change-Id: I24375f290d6c2cde782c43373a07e78ba5d806ff